requirements.txt
requests
streamlit
python-dotenv
langchain-openai
langchain-core
pydantic
